// src/api/types.ts
export interface Paid {
  amount: number;
  currency: 'USD' | 'GBP' | 'EUR';
}